Add foundational guard tests and fix utils path/position conversion bugs.
Expand P0-P2 regression coverage for path/URI helpers, file I/O,
JsonRpc streams, serialization unions, and working-file directory
matching. Fix GetDirName, SplitString, ReadContent, FormatMicroseconds,
and UNC NormalizePath semantics, and make position/offset conversion
count UTF-16 code units including surrogate pairs consistently.

+++ include/LibLsp/lsp/utils.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <algorithm>
 #include <codecvt>
+#include <cstddef>
 #include <functional>
 #include <iterator>
 #include <locale>
@@ -104,7 +105,7 @@ AbsolutePath NormalizePath(std::string const& path, bool ensure_exists = true, b
 int GetOffsetForPosition(lsPosition position, std::string const& content);
 
 // Finds the position for an |offset| in |content|.
-lsPosition GetPositionForOffset(int offset, std::string const& content);
+lsPosition GetPositionForOffset(size_t offset, std::string const& content);
 
 // Utility method to find a position for the given character.
 lsPosition CharPos(std::string const& search, char character, int character_offset = 0);
